Add Gaussian Generator & split into seperate files
data-generator.js now includes Uniform and Gaussian distributions to draw from
